requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
boto3==1.35.37
botocore==1.35.37
click==8.1.7
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==3.0.1
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
